requirements.txt
Django==5.0.4
django-jazzmin==2.6.1
djangorestframework==3.15.1
python-dotenv==1.0.1
pillow==10.3.0
django-mptt==0.16.0